bug fix en importacion

+++ src/main/java/ar/com/marcelomingrone/derechosAutor/estadisticas/batch/DatosCancionMapper.java
@@ -51,7 +51,7 @@ public class DatosCancionMapper implements FieldSetMapper<DatosCancion> {
 			derecho.setModificable(false);
 			datosCancion.setDerecho(derecho);
 			
-			long copyRightShares = fieldSet.readLong(i++);
+			double copyRightShares = fieldSet.readDouble(i++);
 			long unidades = fieldSet.readLong(i++);
 			long externalCode = fieldSet.readLong(i++);
 			double currencyFactor = fieldSet.readDouble(i++);
